Split cluster tests
