+++ herddb-core/src/test/java/herddb/cluster/follower/MultiServerBase.java
@@ -1,0 +1,61 @@
+/*
+ Licensed to Diennea S.r.l. under one
+ or more contributor license agreements. See the NOTICE file
+ distributed with this work for additional information
+ regarding copyright ownership. Diennea S.r.l. licenses this file
+ to you under the Apache License, Version 2.0 (the
+ "License"); you may not use this file except in compliance
+ with the License.  You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing,
+ software distributed under the License is distributed on an
+ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ KIND, either express or implied.  See the License for the
+ specific language governing permissions and limitations
+ under the License.
+
+ */
+
+package herddb.cluster.follower;
+
+import herddb.utils.SystemInstrumentation;
+import herddb.utils.ZKTestEnv;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.rules.TemporaryFolder;
+
+/**
+ * Booting two servers, one table space
+ *
+ * @author enrico.olivelli
+ */
+public class MultiServerBase {
+
+    @Rule
+    public TemporaryFolder folder = new TemporaryFolder();
+
+    protected ZKTestEnv testEnv;
+
+    @Before
+    public void beforeSetup() throws Exception {
+        testEnv = new ZKTestEnv(folder.newFolder().toPath());
+        testEnv.startBookie();
+    }
+
+    @After
+    public void afterTeardown() throws Exception {
+        SystemInstrumentation.clear();
+        if (testEnv != null) {
+            testEnv.close();
+        }
+    }
+
+
+
+
+
+
+}
